+++ tests/test_get_presigned_object_url.rs
@@ -1,0 +1,36 @@
+// MinIO Rust Library for Amazon S3 Compatible Cloud Storage
+// Copyright 2025 MinIO, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+mod common;
+
+use crate::common::{create_bucket_helper, rand_object_name, TestContext};
+use http::Method;
+use minio::s3::args::GetPresignedObjectUrlArgs;
+
+#[tokio::test(flavor = "multi_thread", worker_threads = 10)]
+async fn get_presigned_object_url() {
+    let ctx = TestContext::new_from_env();
+    let (bucket_name, _cleanup) = create_bucket_helper(&ctx).await;
+
+    let object_name = rand_object_name();
+    let resp = ctx
+        .client
+        .get_presigned_object_url(
+            &GetPresignedObjectUrlArgs::new(&bucket_name, &object_name, Method::GET).unwrap(),
+        )
+        .await
+        .unwrap();
+    assert!(resp.url.contains("X-Amz-Signature="));
+}
